requirements.txt
fastapi
uvicorn
langchain-core
langchain-google-genai
langgraph
pinecone-client
langchain-pinecone
pdfplumber
fpdf2
python-dotenv